fix: fixed the query for getting prompt analytics

+++ backend/db_connection.py
@@ -973,8 +973,8 @@ class SupabaseDB:
                         COALESCE(AVG(s.duration), 0) AS avg_latency,
                         COUNT(DISTINCT CASE WHEN s.error_message IS NOT NULL THEN s.trace_id END) AS error_traces,
                         ROUND(
-                            (COUNT(DISTINCT CASE WHEN s.error_message IS NOT NULL THEN s.trace_id END)::NUMERIC
-                            / NULLIF(COUNT(DISTINCT s.trace_id), 0) * 100)::NUMERIC, 2
+                            COUNT(DISTINCT CASE WHEN s.error_message IS NOT NULL THEN s.trace_id END)::NUMERIC
+                            / NULLIF(COUNT(DISTINCT s.trace_id), 0) * 100, 2
                         ) AS error_rate_pct
                     FROM prompt_versions pv
                     LEFT JOIN spans s ON s.prompt_id = pv.prompt_id
@@ -1026,8 +1026,8 @@ class SupabaseDB:
                         COALESCE(AVG(s.duration), 0) AS avg_latency,
                         COUNT(DISTINCT CASE WHEN s.error_message IS NOT NULL THEN s.trace_id END) AS error_traces,
                         ROUND(
-                            (COUNT(DISTINCT CASE WHEN s.error_message IS NOT NULL THEN s.trace_id END)::NUMERIC
-                            / NULLIF(COUNT(DISTINCT s.trace_id), 0) * 100)::NUMERIC, 2
+                            COUNT(DISTINCT CASE WHEN s.error_message IS NOT NULL THEN s.trace_id END)::NUMERIC
+                            / NULLIF(COUNT(DISTINCT s.trace_id), 0) * 100, 2
                         ) AS error_rate_pct
                     FROM prompt_versions pv
                     LEFT JOIN spans s ON s.prompt_id = pv.prompt_id
